backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
sqlalchemy
psycopg[binary]
celery
redis
passlib[bcrypt]
pyjwt
numpy
opencv-python
mediapipe
torch
torchaudio
transformers
sentence-transformers
librosa
soundfile
openai-whisper
openai
ffmpeg-python
Pillow
pytest
livekit-api
livekit-agents
livekit-plugins-openai
livekit-plugins-simli
